# entrega1/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/concert/{concertId}/ticket:
    post:
      tags:
      - ticket-rest-controller
      summary: Confirm purchase of tickets
      description: Confirms the purchase of tickets for a specific concert. Requires
        concert ID to link the tickets with the concert.
      operationId: confirmPurchaseREST
      parameters:
      - name: concertId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: number
        in: query
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketDTO"
        required: true
      responses:
        "201":
          description: Ticket successfully purchased.
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: "Bad Request: Purchase failed."
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: "Unauthorized: User is not authenticated."
          content:
            '*/*':
              schema:
                type: object
...
  /api/tickets/:
    get:
      tags:
      - ticket-rest-controller
      operationId: getUserTickets
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelTicketDTO"
...
  /api/tickets/{id}:
    delete:
      tags:
      - ticket-rest-controller
      summary: Delete a ticket
      description: "Deletes a specific ticket from the system using its ID. If the\
        \ ticket does not exist or an error occurs during deletion, it will be handled\
        \ by a global exception handler."
      operationId: deleteTicketREST
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Ticket successfully deleted.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TicketDTO"
        "400":
          description: "Bad Request: Ticket could not be deleted."
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TicketDTO"
        "404":
          description: Ticket not found with the provided ID.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TicketDTO"
        "500":
          description: Server error during ticket deletion.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TicketDTO"
components:
  schemas:
    ShowUserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        country:
          type: string
        email:
          type: string
    Artist:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        popularityIndex:
          type: integer
          format: int64
        sessionCreated:
          type: string
          format: date-time
        photoLink:
          type: string
        hasPage:
          type: boolean
        mainInfo:
          type: string
        extendedInfo:
          type: string
        bestAlbumSpotifyLink:
          type: string
        latestAlbumSpotifyLink:
          type: string
        videoLink:
          type: string
    ConcertDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        artist:
          $ref: "#/components/schemas/Artist"
        name:
          type: string
        info:
          type: string
        date:
          type: string
          format: date-time
        place:
          type: string
        price:
          type: number
          format: float
        westStandsNumber:
          type: integer
          format: int32
        eastStandsNumber:
          type: integer
          format: int32
        northStandsNumber:
          type: integer
          format: int32
        generalAdmissionNumber:
          type: integer
          format: int32
        formattedTime:
          type: string
...
    TicketDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        zone:
          type: string
        price:
          type: number
          format: float
        ticketUser:
          $ref: "#/components/schemas/ShowUserDTO"
        concert:
          $ref: "#/components/schemas/ConcertDTO"
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
...
    PagedModelTicketDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/TicketDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"
...